requirements/dev.txt
-r base.txt

# Debug toolbar
django-debug-toolbar==1.4

# Flake8
pep8==1.5.7
flake8==2.5.0

# Flake8 plugins
flake8-coding==1.1.1
flake8-copyright==0.1
flake8-future-import==0.3.2
flake8-import-order==0.6.1
flake8-print==2.0.1
pep8-naming==0.3.3
flake8-quotes==0.1.1
pep257==0.6.0
flake8-pep257==1.0.2
